Add the POC of ChainSyncer
The consensus nodes use ChainSyncer to ensure they have the same view.
ChainSyncer manages the states and requests ChainSyncerClient to fetch
data. The caller of ChainSyncer is responsible to update ChainSyncer's
states.

Note that this commit does not fulfill a complete ChainSyncer. The goal
of this commit is to define the responsibility between ChainSyncer and
ChainSyncerClient and provides a naive design to demonstrate the
concepts.

The usage of ChainSyncer will be more clear in the next commit.

Testing: go test -v -race -count=1 thunder2/consensus